+++ CHANGELOG.md
@@ -3,8 +3,10 @@
 ## Unreleased
 
 - New features
-    - #81: Support for specifying process-compose global CLI options
-      - **Breaking change**: Old options `httpServer` and `tui` were removed; users should use the new `cli` module to set all process-compose cli arguments and options.
+    - #81, #84: Support for specifying process-compose global CLI options
+      - **Breaking changes**:
+        - `preHook` and `postHook` are now inside `cli` module.
+        - Old options `httpServer` and `tui` were removed; users should use the new `cli` module to set all process-compose cli arguments and options.
     - ~~#18: Add `testScript` option for adding flake checks based on nixosTest library.~~
     - #39: Allow `test` process to act as a test, which then gets run as part of flake checks.
     - #55: Add `lib` flake output - library of useful functions